+++ USER/HARDWARE/BSP_pressure_sensor.h
@@ -1,3 +1,4 @@
+
 #ifndef  	_BSP_PRESSURE_SENSOR_H_
 #define 	_BSP_PRESSURE_SENSOR_H_
 
@@ -8,8 +9,8 @@
 /*****************************************************/
 //压力传感器参数配置
 
-#define		IN_VOLTAGE_MAX				3000		//单位mv //单片机采集最大电压
-#define 	PRESSURE_SENSOR_SPAN_MAX	300		//最大量程 重量	牛顿 N
+#define		IN_VOLTAGE_MAX						3000		//单位mv //单片机采集最大电压
+#define 	PRESSURE_SENSOR_SPAN_MAX	300			//最大量程 重量	牛顿 N
 
 /*****************************************************/
 
@@ -25,6 +26,6 @@
 
 void bsp_Pressure_Sensor_Init(void);							//初始化
 float Pressure_Sensor_Get_Voltage(void);						//获取电压
-int bsp_Pressure_Sensor_Get_Newton(void);							//获取压力值，单位牛顿
+float bsp_Pressure_Sensor_Get_Newton(void);							//获取压力值，单位牛顿
 
 #endif
